services/ocr/requirements.txt
# GLM-OCR Dependencies
torch==2.1.2
transformers==4.37.2
accelerate==0.26.1
pillow==10.2.0
opencv-python==4.9.0.80

# API Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Utilities
numpy==1.26.3
httpx==0.26.0
python-dotenv==1.0.0
